requirements.txt
flask==2.3.2
pandas==1.5.3
torch==2.0.1
open-clip-torch==2.16.0
numpy==1.24.3
scipy==1.10.1
pillow==9.5.0
